requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
openai>=1.40.0
langgraph>=0.2.6
pillow>=10.4.0
python-docx>=1.1.2
